+++ app/controllers/services_controller.rb
@@ -55,13 +55,26 @@ class ServicesController < ApplicationController
   end
 
   def hidden_user_explanation
-    reason = render_to_string(
+    prominence_reason = render_to_string(
       partial: "admin_request/hidden_user_explanation/#{params[:message]}",
-      formats: [:text]
+      formats: [:text],
+      locals: {
+        prominence_reason: true,
+        explanation: false
+      }
+    )
+
+    explanation = render_to_string(
+      partial: "admin_request/hidden_user_explanation/#{params[:message]}",
+      formats: [:text],
+      locals: {
+        prominence_reason: false,
+        explanation: true
+      }
     )
 
     render json: {
-      prominence_reason: reason,
+      prominence_reason: prominence_reason,
       explanation: render_to_string(
         template: "admin_request/hidden_user_explanation",
         formats: [:text],
@@ -69,7 +82,7 @@ class ServicesController < ApplicationController
         locals: {
           name_to: @info_request.user_name.html_safe,
           info_request: @info_request,
-          reason: reason,
+          explanation: explanation,
           info_request_url: request_url(@info_request),
           site_name: site_name.html_safe
         }